refactor(advisor-book): isolate request dependencies (#500)
